refactor: move types around
